modify UI code
