+++ Seminar/SOPT_iOS_SEMINAR.xcodeproj/xcuserdata/j.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -8,7 +8,7 @@
          BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
          <BreakpointContent
             uuid = "0ABAF2BA-516A-41BA-8157-FDA1A4BD11DF"
-            shouldBeEnabled = "No"
+            shouldBeEnabled = "Yes"
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "SOPT_iOS_SEMINAR/FirstViewController_button.swift"
@@ -24,7 +24,7 @@
          BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
          <BreakpointContent
             uuid = "39FE9E6B-B47D-4CD0-AF48-C1BBB2F4AB99"
-            shouldBeEnabled = "No"
+            shouldBeEnabled = "Yes"
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "SOPT_iOS_SEMINAR/FirstViewController_button.swift"
