+++ Seminar/SOPT_iOS_SEMINAR/Assets.xcassets/assignment3/poster_0.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "image 21.pdf",
+      "filename" : "denti_da_squalo_ver2.jpg",
       "idiom" : "universal"
     }
   ],
