+++ Seminar/SOPT_iOS_SEMINAR/Assets.xcassets/assignment3/poster_5.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "image 47.pdf",
+      "filename" : "y_todos_arderan_ver2.jpg",
       "idiom" : "universal"
     }
   ],
