+++ Seminar/assignment3/MainPage/MyCell.swift
@@ -11,18 +11,19 @@ import SnapKit
 import Kingfisher
 
 final class MyCell: UICollectionViewCell {
+    
     static let id = "MyCell"
     
     // MARK: UI
     private let imageView = UIImageView().then {
         $0.image = UIImage(named: "Image")
-        $0.backgroundColor = .tvingWhite
+//        $0.backgroundColor = .tvingWhite
         $0.contentMode = .scaleAspectFill
         $0.translatesAutoresizingMaskIntoConstraints = false
         $0.makeCornerRound(radius: 5)
     }
     private lazy var label = UILabel().then {
-        $0.text = "우리가 사랑했던 모든 것"
+        $0.text = ""
         
         $0.font = UIFont(name: "Pretendard-Medium", size: 12)
         $0.textColor = .tvingWhite
@@ -39,11 +40,11 @@ final class MyCell: UICollectionViewCell {
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        
         self.contentView.addSubview(self.imageView)
         self.contentView.addSubview(self.label)
         imageView.snp.makeConstraints {
-            $0.top.equalToSuperview()
+            $0.top.equalToSuperview().inset(10)
+//                .inset(-10)
             $0.leading.equalToSuperview()
             $0.height.equalTo(150)
             $0.width.equalTo(115)
