+++ Seminar/assignment3/MainPage/TableViewCell2.swift
@@ -14,12 +14,13 @@ import Then
 class TableViewCell2: UITableViewCell {
     static let identifier = "section1"
     let fullSize = UIScreen.main.bounds
-    
-    
     private let flowLayout = UICollectionViewFlowLayout().then {
+        $0.sectionInset = UIEdgeInsets(top: 0, left: 10, bottom: 0, right: 10)
         $0.scrollDirection = .horizontal
         $0.minimumInteritemSpacing = 8.0
-        $0.itemSize = CGSize(width: 115, height: 150)
+        $0.itemSize = CGSize(width: 123, height: 190)
+        $0.minimumLineSpacing = 0.0 // 상하간격
+//        $0.minimumInteritemSpacing = 8
     }
     
     var movieList: [Result] = [] {
@@ -32,10 +33,8 @@ class TableViewCell2: UITableViewCell {
         
         $0.isScrollEnabled = true
         $0.showsHorizontalScrollIndicator = false
-        $0.showsVerticalScrollIndicator = true
         $0.contentInset = .zero
-        $0.backgroundColor = .systemPink
-        $0.clipsToBounds = true
+        $0.backgroundColor = .tvingBlack
         $0.register(MyCell.self, forCellWithReuseIdentifier: "MyCell")
         $0.translatesAutoresizingMaskIntoConstraints = false
     }
@@ -43,6 +42,7 @@ class TableViewCell2: UITableViewCell {
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         
+        self.contentView.backgroundColor = .tvingBlack
         setSubviews()
         setLayouts()
         test()
@@ -61,11 +61,12 @@ class TableViewCell2: UITableViewCell {
         collectionView.snp.makeConstraints {
             $0.top.equalToSuperview()
             $0.leading.equalToSuperview()
-            $0.height.equalTo(200)
+            $0.height.equalTo(170)
             $0.trailing.equalToSuperview()
         }
         
         self.collectionView.dataSource = self
+        self.selectionStyle = .none
     }
     
     
@@ -100,57 +101,3 @@ extension TableViewCell2 {
 }
 
 
-
-extension TableViewCell2: UITableViewDataSource {
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return 2
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return list[section].items.count
-        
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        switch indexPath.section {
-        case 0:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: MyPage_TableViewCell.cellIdentifier, for: indexPath) as? MyPage_TableViewCell else { return UITableViewCell() }
-            cell.configureCell(list[0].items[indexPath.row])
-            return cell
-
-        case 1:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: MyPage_TableViewCell.cellIdentifier, for: indexPath) as? MyPage_TableViewCell else { return UITableViewCell() }
-            cell.configureCell(list[1].items[indexPath.row])
-            return cell
-        default: return UITableViewCell()
-            
-        }
-    }
-    
-
-    // MARK : Header, Footer
-    
-    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        switch section {
-        case 0: return tableView.dequeueReusableHeaderFooterView(withIdentifier: "SeparatorTableViewFooter")
-        case 1: return tableView.dequeueReusableHeaderFooterView(withIdentifier: "LogOutButtonFooter")
-        default: return nil
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        guard section == 0, let headerCell = tableView.dequeueReusableHeaderFooterView(withIdentifier: "MyTableViewHeader") as? MyTableViewHeader else { return nil }
-        
-        let tapGesture = UITapGestureRecognizer()
-        headerCell.addGestureRecognizer(tapGesture)
-        return headerCell
-    }
-    
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return section == 0 ? 310 : 0
-    }
-    
-    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        return section == 1 ? 100 : 33
-    }
-}
