+++ Seminar/assignment4/Movie/MovieResponse.swift
@@ -1,0 +1,64 @@
+//
+//  MovieResponse.swift
+//  SOPT_iOS_SEMINAR
+//
+//  Created by Minjoo Kim on 2023/05/09.
+//
+
+import Foundation
+
+
+// MARK: - Welcome
+struct MovieResponse: Codable {
+    let dates: Dates
+    let page: Int
+    let results: [Result]
+    let totalPages, totalResults: Int
+
+    enum CodingKeys: String, CodingKey {
+        case dates, page, results
+        case totalPages = "total_pages"
+        case totalResults = "total_results"
+    }
+}
+
+// MARK: - Dates
+struct Dates: Codable {
+    let maximum, minimum: String
+}
+
+// MARK: - Result
+struct Result: Codable {
+    let adult: Bool
+    let backdropPath: String
+    let genreIDS: [Int]
+    let id: Int
+    let originalLanguage: OriginalLanguage
+    let originalTitle, overview: String
+    let popularity: Double
+    let posterPath, releaseDate, title: String
+    let video: Bool
+    let voteAverage: Double
+    let voteCount: Int
+
+    enum CodingKeys: String, CodingKey {
+        case adult
+        case backdropPath = "backdrop_path"
+        case genreIDS = "genre_ids"
+        case id
+        case originalLanguage = "original_language"
+        case originalTitle = "original_title"
+        case overview, popularity
+        case posterPath = "poster_path"
+        case releaseDate = "release_date"
+        case title, video
+        case voteAverage = "vote_average"
+        case voteCount = "vote_count"
+    }
+}
+
+enum OriginalLanguage: String, Codable {
+    case en = "en"
+    case es = "es"
+    case fi = "fi"
+}
